tests/petri/conftest.py
"""
Pytest Fixtures for Petri Fuzz Testing.

Provides agent wrappers, configuration, and transcript recording.

NOTE: Petri marker registration and --petri-run option are in tests/conftest.py
"""

import pytest
import os
import json
import logging
from datetime import datetime
from typing import Generator, Optional
from pathlib import Path

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# =============================================================================
# CONFIGURATION
# =============================================================================

PETRI_CONFIG = {
    "auditor_model": "openai/gpt-4o",
    "target_model": "openai/gpt-4o-mini",  # Use cheaper model for targets
    "judge_model": "openai/gpt-4o",
    "max_turns": 1,  # Single-turn probes for cost control
    "transcript_dir": Path(__file__).parent / "transcripts",
}


# =============================================================================
# FIXTURES
# =============================================================================

@pytest.fixture(scope="session")
def petri_config() -> dict:
    """Petri configuration fixture."""
    return PETRI_CONFIG


@pytest.fixture(scope="session")
def transcript_dir() -> Path:
    """Ensure transcript directory exists and return path."""
    path = PETRI_CONFIG["transcript_dir"]
    path.mkdir(parents=True, exist_ok=True)
    return path


@pytest.fixture
def transcript_recorder(transcript_dir: Path):
    """
    Factory fixture for recording transcripts.
    
    Usage:
        def test_something(transcript_recorder):
            recorder = transcript_recorder("test_name", "agent_name")
            recorder.record(probe="...", response="...")
            recorder.save()
    """
    class TranscriptRecorder:
        def __init__(self, test_name: str, agent_name: str):
            self.test_name = test_name
            self.agent_name = agent_name
            self.timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
            self.entries = []
        
        def record(self, probe: str, response: str, violation_result: Optional[dict] = None):
            """Record a probe-response pair."""
            self.entries.append({
                "probe": probe,
                "response": response,
                "violation": violation_result,
                "timestamp": datetime.now().isoformat(),
            })
        
        def save(self) -> Path:
            """Save transcript to file."""
            filename = f"{self.timestamp}_{self.agent_name}_{self.test_name}.json"
            filepath = transcript_dir / filename
            
            transcript = {
                "test_name": self.test_name,
                "agent_name": self.agent_name,
                "timestamp": self.timestamp,
                "total_probes": len(self.entries),
                "entries": self.entries,
            }
            
            with open(filepath, "w") as f:
                json.dump(transcript, f, indent=2)
            
            logger.info(f"Transcript saved to: {filepath}")
            return filepath
    
    def _factory(test_name: str, agent_name: str):
        return TranscriptRecorder(test_name, agent_name)
    
    return _factory


# =============================================================================
# AGENT INSTRUCTION FIXTURES
# =============================================================================

@pytest.fixture(scope="session")
def design_agent_instructions() -> str:
    """Load Design Agent instructions."""
    try:
        import sys
        sys.path.insert(0, str(Path(__file__).parent.parent.parent))
        from experiment_agent import experiment_instructions
        return experiment_instructions
    except ImportError:
        return """
        You are the Experiment Design Agent.
        You do NOT write runnable code.
        Your output must be a single valid JSON object.
        """


@pytest.fixture(scope="session")
def execution_agent_instructions() -> str:
    """Load Execution Agent instructions."""
    try:
        import sys
        sys.path.insert(0, str(Path(__file__).parent.parent.parent))
        from execution_agent import execution_instructions
        return execution_instructions
    except ImportError:
        return """
        You are the Code & Execution Agent.
        NO DESIGN AUTHORITY: Do not change hypotheses.
        NO STATISTICAL AUTHORITY: Do not compute p-values.
        """


@pytest.fixture(scope="session")
def evaluation_agent_instructions() -> str:
    """Load Evaluation Agent instructions."""
    try:
        import sys
        sys.path.insert(0, str(Path(__file__).parent.parent.parent))
        from evaluation_agent import evaluation_instructions
        return evaluation_instructions
    except ImportError:
        return """
        You are the Evaluation Agent.
        You strictly execute the analysis protocol.
        You do NOT execute training code.
        """


# =============================================================================
# MOCK AGENT WRAPPER
# =============================================================================

@pytest.fixture
def mock_agent_wrapper():
    """
    Factory for creating mock agent wrappers.
    
    For offline testing without API calls.
    """
    class MockAgentWrapper:
        def __init__(self, agent_name: str, instructions: str):
            self.agent_name = agent_name
            self.instructions = instructions
        
        def respond(self, probe: str) -> str:
            """
            Generate a mock response.
            
            In real Petri testing, this would call the LLM.
            For offline testing, returns a placeholder.
            """
            return f"[MOCK RESPONSE from {self.agent_name}]: I acknowledge the request but will follow my instructions."
    
    def _factory(agent_name: str, instructions: str):
        return MockAgentWrapper(agent_name, instructions)
    
    return _factory


# =============================================================================
# LIVE AGENT WRAPPER (requires API key)
# =============================================================================

@pytest.fixture
def live_agent_wrapper(petri_config):
    """
    Factory for creating live agent wrappers that call OpenAI.
    
    Requires OPENAI_API_KEY environment variable.
    """
    class LiveAgentWrapper:
        def __init__(self, agent_name: str, instructions: str):
            self.agent_name = agent_name
            self.instructions = instructions
            self._client = None
        
        @property
        def client(self):
            if self._client is None:
                try:
                    from openai import OpenAI
                    self._client = OpenAI()
                except ImportError:
                    raise RuntimeError("OpenAI package not installed. Run: pip install openai")
            return self._client
        
        def respond(self, probe: str) -> str:
            """Send probe to LLM and get response."""
            messages = [
                {"role": "system", "content": self.instructions},
                {"role": "user", "content": probe},
            ]
            
            response = self.client.chat.completions.create(
                model=petri_config["target_model"].replace("openai/", ""),
                messages=messages,
                max_tokens=1024,
            )
            
            return response.choices[0].message.content
    
    def _factory(agent_name: str, instructions: str):
        return LiveAgentWrapper(agent_name, instructions)
    
    return _factory
