requirements.txt
huggingface-hub==0.36.0
beautifulsoup4>=4.12.0
numpy==2.3.5
openai-agents==0.6.3
pandas==2.3.3
pypdf==6.4.2
python-dotenv==1.2.1
requests==2.32.5
scikit-learn==1.8.0
scipy==1.16.3
statsmodels==0.14.6
supabase==2.3.5
tavily-python==0.7.16
torch==2.6.0+cu124
torchvision==0.21.0+cu124
matplotlib>=3.8.0

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pyyaml>=6.0.0

# Optional: Petri for instruction fuzz testing (run: pip install git+...)
# petri @ git+https://github.com/safety-research/petri.git
